+++ app/Http/Controllers/Api/OfferController.php
@@ -9,6 +9,12 @@ use App\Http\Requests\Api\OfferExternalUpdateRequest;
 use Illuminate\Http\Request;
 use App\Models\Offer;
 use App\Models\Institution;
+use App\Models\Competence;
+use App\Models\Meta;
+use App\Models\Language;
+use App\Models\Huboffer;
+use App\Models\Offertype;
+use App\Models\Timestamp;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
@@ -23,14 +29,17 @@ class OfferController extends Controller
     public function index()
     {
         $offers = Offer::all();
-        $offers = $offers->sortByDesc("sort_flag");
         $offers = $offers->values()->all();
+        $sort = array();
+        $output = array();
+        foreach ( $offers as $offer ) {
+            $output[] = $this->restructureJsonOutput($offer);
+            $sort[$offer->id] = $offer->huboffer->sort_flag;
+        }
+        array_multisort($sort, SORT_DESC, $output);
 
-        #todo
-
-        return response()->json($offers, 200);
+        return response()->json($output, 200);
     }
-
 
     /**
      * Store a newly created resource in storage.
@@ -40,17 +49,14 @@ class OfferController extends Controller
      */
     public function store(OfferStoreRequest $request)
     {
-
-        $validatedData = $request->validated();
+        $validatedData = $this->validateRedundantInput( $request->validated() );
 
         $offer = Offer::create($validatedData);
         $offer->save();
-        # The response would only contain the entered data. Create new complete object for response
-        $id = $offer->id;
-        $offer = new Offer;
-        $offer = $offer->find($id);
+        $this->saveRelatedData( $offer, $validatedData );
+        $offer = Offer::find($offer->id);
 
-        return response()->json($offer, 201);
+        return response()->json($this->restructureJsonOutput($offer), 201);
     }
 
     /**
@@ -68,40 +74,44 @@ class OfferController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \App\Http\Requests\Api\OfferUpdateRequest  $request
-     * @param  \App\Models\Offer  $offer
+     * @param  \App\Models\Offer $offer
      * @return \Illuminate\Http\Response
      */
-    public function update(OfferUpdateRequest $request, int $id)
+    public function update(OfferUpdateRequest $request, Offer $offer)
     {
-        $offer = Offer::find($id);
-        $validatedData = $request->validated();
+        $validatedData = $this->validateRedundantInput( $request->validated() );
 
         $offer->fill($validatedData);
         $offer->save();
-        return response()->json($offer, 201);
+        $this->saveRelatedData( $offer, $validatedData );
+        $offer = Offer::find($offer->id);
+
+        return response()->json($this->restructureJsonOutput($offer), 201);
     }
 
     /**
      * Update the specified resource by given external id in storage.
      *
-     * @param  \App\Http\Requests\Api\OfferUpdateRequest  $request
+     * @param  \App\Http\Requests\Api\OfferExternalUpdateRequest  $request
      * @param  String $externalId
      * @return \Illuminate\Http\Response
      */
     public function updateByExternalId(OfferExternalUpdateRequest $request, Institution $institution, String $externalId)
     {
         $offer = Offer::where(["institution_id" => $institution->id, "externalId" => $externalId ])->firstOrFail();
-        $validatedData = $request->validated();
+        $validatedData = $this->validateRedundantInput( $request->validated() );
+        unset( $validatedData["institution_id"], $validatedData["externalId"], $validatedData["sort_flag"], $validatedData["keywords"], $validatedData["visible"] );
         $offer->fill($validatedData);
         $offer->save();
+        $this->saveRelatedData( $offer, $validatedData );
 
-        return response()->json($offer, 201);
+        return response()->json($this->restructureJsonOutput($offer), 201);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Offer  $offer
+     * @param  \App\Models\Offer $offer
      * @return \Illuminate\Http\Response
      */
     public function destroy(Offer $offer)
@@ -110,43 +120,135 @@ class OfferController extends Controller
         return response(null, 204);
     }
 
-    public function restructureJsonOutput( Offer $offer ) {
-        $arr = $offer->toArray();
-        $arr["language"] = $arr["language"]["identifier"];
-        $arr["sort_flag"] = $arr["huboffer"]["sort_flag"];
-        $arr["keywords"] = $arr["huboffer"]["keywords"];
-        $arr["visible"] = $arr["huboffer"]["visible"];
-        $arr["type"] = $arr["offertype"]["identifier"];
-        $arr["meta"] = [];
-        foreach ( $arr["metas"] as $meta ) {
-            $arr["meta"][$meta["description"]] = $meta["pivot"]["value"];
+    /**
+     * Remodel the output of an offer
+     *
+     * @param  \App\Models\Offer $offer
+     * @return Array $ret
+     */
+    private function restructureJsonOutput( Offer $offer ) {
+
+        $ret = $offer->toArray();
+
+        $ret["language"] = $ret["language"]["identifier"];
+
+        if ( is_array( $ret["huboffer"] ) ) {
+            $ret["sort_flag"] = $ret["huboffer"]["sort_flag"];
+            $ret["keywords"] = $ret["huboffer"]["keywords"];
+            $ret["visible"] = $ret["huboffer"]["visible"];
         }
-        $arr["competence_tech"] = 0;
-        $arr["competence_digital"] = 0;
-        $arr["competence_classic"] = 0;
-        foreach ( $arr["competences"] as $competence ) {
-            $arr["competence_".$competence["title"]] = 1;
+
+        $ret["type"] = $ret["offertype"]["identifier"];
+
+        $ret["meta"] = [];
+        foreach ( $ret["metas"] as $meta ) {
+            $ret["meta"][$meta["description"]] = $meta["pivot"]["value"];
         }
-        $arr["relations"] = [];
-        foreach ( $arr["original_relations"] as $relation ) {
-            $arr["relations"][] = $relation["id"];
+
+        $ret["competence_tech"] = 0;
+        $ret["competence_digital"] = 0;
+        $ret["competence_classic"] = 0;
+        foreach ( $ret["competences"] as $competence ) {
+            $ret["competence_".$competence["title"]] = 1;
+        }
+
+        $ret["relations"] = [];
+        foreach ( $ret["original_relations"] as $relation ) {
+            $ret["relations"][] = $relation["id"];
         }
 
         unset(
-            $arr["competences"],
-            $arr["metas"],
-            $arr["huboffer"],
-            $arr["offertype"],
-            $arr["original_relations"],
-            $arr["timestamps"]["id"],
-            $arr["timestamps"]["id"],
-            $arr["timestamps"]["created_at"],
-            $arr["timestamps"]["updated_at"],
-            $arr["institution"]["created_at"],
-            $arr["institution"]["updated_at"],
-            $arr["institution"]["json_url"]
+            $ret["competences"],
+            $ret["metas"],
+            $ret["huboffer"],
+            $ret["offertype"],
+            $ret["original_relations"],
+            $ret["timestamps"]["id"],
+            $ret["timestamps"]["id"],
+            $ret["timestamps"]["created_at"],
+            $ret["timestamps"]["updated_at"],
+            $ret["timestamps"]["offer_id"],
+            $ret["institution"]["created_at"],
+            $ret["institution"]["updated_at"],
+            $ret["institution"]["json_url"]
         );
 
-        return $arr;
+        return $ret;
     }
+
+    /**
+     * Save an offer's data in other tables
+     *
+     * @param  \App\Models\Offer $offer
+     * @param  Array $validatedData
+     */
+    private function saveRelatedData( Offer $offer, Array $validatedData ) {
+
+        # Sync pivot tables
+        $competences = Competence::all();
+        $competence_sync = array();
+        foreach ( $competences as $c ) {
+            if ( \key_exists( "competence_".$c->title, $validatedData ) && $validatedData["competence_".$c->title] == true ) {
+                $competence_sync[] = $c->id;
+            }
+        }
+        $offer->competences()->sync($competence_sync);
+
+        $metas = Meta::all();
+        $meta_sync = array();
+        foreach ( $metas as $m ) {
+            if ( \key_exists( $m->description, $validatedData ) && !empty ( $validatedData[$m->description] ) ) {
+                $meta_sync[ $m->id ] = [ "value" => $validatedData[$m->description] ];
+            }
+        }
+        $offer->metas()->sync($meta_sync);
+
+        # Fill other related tables
+        $hubOffer = Huboffer::where([ "offer_id" => $offer->id ])->first();
+        if ( is_object( $hubOffer ) ) {
+            $hubOffer->fill($validatedData);
+            $hubOffer->save();
+        }  else {
+            $validatedData["offer_id"] = $offer->id;
+            $hubOffer = Huboffer::create($validatedData);
+        }
+
+        $timestamp = Timestamp::where([ "offer_id" => $offer->id ])->first();
+        if ( is_object( $timestamp ) ) {
+            $timestamp->fill($validatedData);
+            $timestamp->save();
+        }  else {
+            $validatedData["offer_id"] = $offer->id;
+            $timestamp = Timestamp::create($validatedData);
+        }
+
+    }
+
+    /**
+     * Support for using API readable parameterd
+     * 'type' instead of 'offertype_id'
+     * 'language' instead of 'language_id'
+     *
+     * @param  Array $validatedData
+     * @return  Array $validatedData
+     */
+    private function validateRedundantInput( Array $validatedData ) {
+
+        if ( !key_exists ( "offertype_id", $validatedData ) && !empty ( $validatedData["type"] ) ) {
+            $offertype = Offertype::where([ "identifier" => $validatedData["type"] ])->first();
+            if ( is_object( $offertype ) ) {
+                $validatedData["offertype_id"] = $offertype->id;
+            }
+        }
+
+        if ( !key_exists ( "language_id", $validatedData ) && !empty ( $validatedData["language"] ) ) {
+            $language = Language::where([ "identifier" => $validatedData["language"] ])->first();
+            if ( is_object( $language ) ) {
+                $validatedData["language_id"] = $language->id;
+            }
+        }
+        return $validatedData;
+
+    }
+
 }