+++ app/Http/Requests/Api/OfferUpdateRequest.php
@@ -4,7 +4,6 @@ namespace App\Http\Requests\Api;
 
 use Illuminate\Foundation\Http\FormRequest;
 use Illuminate\Validation\Rule;
-use App\Models\Offer;
 
 class OfferUpdateRequest extends FormRequest
 {
@@ -26,49 +25,55 @@ class OfferUpdateRequest extends FormRequest
     public function rules()
     {
         return [
-            'title' => 'required|max:255',
-            'offertype_id' => 'integer', #RENAMED
+            'title' => 'max:255',
+            'offertype_id' => 'integer',
             'description' => 'nullable|string',
             'image_path' => 'nullable|string',
             'institution_id' => 'integer',
             'subtitle' => 'nullable|string',
-            'language_id' => 'nullable|integer', #RENAMED
+            'language_id' => 'integer',
             'hashtag' => 'nullable|string',
             'author' => 'nullable|string',
             'target_group' => 'nullable|string',
             'url' => 'nullable|url',
-            'externalId' => ['nullable','string', Rule::unique('offers')->where(function ($query) { #RENAMED
-                // custom rule: if the ext_id wasn't new the data would not update
+            'externalId' => ['nullable','string', Rule::unique('offers')->where(function ($query) {
+                // custom rule: if the externalId wasn't new the data would not update
                 // so here we check if the existing data has the same id (which is ok)
                 // id must be in the end of path string.
                 $pathInfo = static::segments($this->getPathInfo());
                 $id = end($pathInfo);
+                #todo custom error message
                 return $query->where(['externalId' => $this->get('externalId'), 'institution_id' => $this->get('institution_id')])->whereNotIn("id", [$id]);
             })],
 
-            #COMPETENCES
+            #COMPETENCES table
             'competence_tech' => 'nullable|boolean',
-            #'competence_digital' => 'nullable|boolean',
-            #'competence_classic' => 'nullable|boolean',
+            'competence_digital' => 'nullable|boolean',
+            'competence_classic' => 'nullable|boolean',
 
-            #METAS
-            #'ects' => 'nullable|integer',
-            #'time_requirement' => 'nullable|string',
-            #'sponsor' => 'nullable|string',
-            #'exam' => 'nullable|string',
-            #'requirements' => 'nullable|string',
-            #'niveau' => 'nullable|string',
+            #METAS table
+            'ects' => 'nullable|integer',
+            'time_requirement' => 'nullable|string',
+            'sponsor' => 'nullable|string',
+            'exam' => 'nullable|string',
+            'requirements' => 'nullable|string',
+            'niveau' => 'nullable|string',
 
-            #HUB_OFFERS
-            #'sort_flag' => 'nullable|integer',
+            #HUBOFFERS table
+            'sort_flag' => 'nullable|integer',
+            'keywords' => 'nullable|string',
+            'visible' => 'nullable|boolean',
 
-            #TIMESTAMPS
-            #'executed_from' => 'date',
-            #'executed_until' => 'nullable|date',
-            #'listed_from' => 'date',
-            #'listed_until' => 'nullable|date',
-            #'active' => 'nullable|boolean',
+            #TIMESTAMPS table
+            'executed_from' => 'date',
+            'executed_until' => 'nullable|date',
+            'listed_from' => 'date',
+            'listed_until' => 'nullable|date',
+            'active' => 'nullable|boolean',
 
+            #Backwards compatibility
+            'type' => 'string',
+            'language' => 'string',
         ];
     }
 }