+++ Changelog.md
@@ -2,6 +2,13 @@
 
 ## [Unreleased]
 
+## 2021-03-15/dk
+### Added
+- Offer_relations table
+- Migrations for DB redesign
+### Changed
+- OfferController and -Requests regarding redesign
+
 ## 2021-02-02/ml
 ### Added
 - BotConfigController.php, reads static JSON File for Bot-Konfiguration (Urls in config/bot.php)